+++ include/util/flops.h
@@ -1,17 +1,15 @@
 #ifndef FLOPS_H
 #define FLOPS_H
 
-#include <stdint.h>
-
 class Flops
 {
 public:
-    uint64_t timestamp;
+    int timestamp;
     float gpu_fp32active;
-    uint64_t duration;
-    Flops(const uint64_t &timestamp,
+    int duration;
+    Flops(const int &timestamp,
           const float &gpu_fp32active,
-          const uint64_t &duration);
+          const int &duration);
     ~Flops();
 };
